Factor out base StreamBuffer from MemoryStreamBuffer
